# Statically define trace observability attributes