--- a/sdk/trace/span.go
+++ b/sdk/trace/span.go
@@ -21,7 +21,6 @@ import (
 	"go.opentelemetry.io/otel/sdk/instrumentation"
 	"go.opentelemetry.io/otel/sdk/resource"
 	semconv "go.opentelemetry.io/otel/semconv/v1.36.0"
-	"go.opentelemetry.io/otel/semconv/v1.36.0/otelconv"
 	"go.opentelemetry.io/otel/trace"
 	"go.opentelemetry.io/otel/trace/embedded"
 )
@@ -499,20 +498,11 @@ func (s *recordingSpan) End(options ...trace.SpanEndOption) {
 
 	if s.tracer.selfObservabilityEnabled {
 		defer func() {
-			// Determine the sampling result and create the corresponding attribute.
-			var attrSamplingResult attribute.KeyValue
-			if s.spanContext.IsSampled() {
-				attrSamplingResult = s.tracer.spanLiveMetric.AttrSpanSamplingResult(
-					otelconv.SpanSamplingResultRecordAndSample,
-				)
-			} else {
-				attrSamplingResult = s.tracer.spanLiveMetric.AttrSpanSamplingResult(otelconv.SpanSamplingResultRecordOnly)
-			}
-
 			// Add the span to the context to ensure the metric is recorded
 			// with the correct span context.
 			ctx := trace.ContextWithSpan(context.Background(), s)
-			s.tracer.spanLiveMetric.Add(ctx, -1, attrSamplingResult)
+			set := spanLiveSet(s.spanContext.IsSampled())
+			s.tracer.spanLiveMetric.AddSet(ctx, -1, set)
 		}()
 	}